update optargorder for Go 1.17
Makes two changes related the register ABI introduced in Go 1.17

1. No longer sorts formal parameters by their address

It doesn't make sense to sort formal parameters by address anymore because
most of them will be stored in registers. Furthermore the DwARF
specification already mandates that they appear in the same order as they
are declared in the source code (DWARFv4 section 3.3.4, page 55).

2. Check location correctness at entrypoint and after the prologue

Performs location vetting at two PC addresses: the function entry point and
the first instruction after the prologue.
The check at entry point can be stricter because we know that register
allocation must conform to some register ABI invariants.
It is interesting to check both addresses because the entry point is
interesting to function call injection but the first instruction after
prologue_end is what users of Delve will see most of the time.